% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_helpers.R
\name{adjust_beta_oneval}
\alias{adjust_beta_oneval}
\title{Adjust a beta coefficient for regression dilution}
\usage{
adjust_beta_oneval(beta, lambda)
}
\arguments{
\item{beta}{Coefficient from epidemiological model to be adjusted}

\item{lambda}{Regression dilution coefficient}
}
\value{
Regression-dilution adjusted beta coefficient
}
\description{
Adjusts a beta coefficient for regression dilution using a calculated value of lambda.
}
