% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_from_model.R
\name{adjust_betas}
\alias{adjust_betas}
\title{Adjust model coefficients for regression dilution}
\usage{
adjust_betas(models, coef_name, lambdas, var_lambdas, hr_scale = TRUE)
}
\arguments{
\item{models}{Models}

\item{coef_name}{Name of coefficient to adjust}

\item{lambdas}{Values of regression dilution coefficient}

\item{var_lambdas}{Variance of regression dilution coefficient estimates}

\item{hr_scale}{If using a Cox model, also return coefficients on HR scale?}
}
\value{
Data frame with original and adjusted coefficient estimates
}
\description{
Adjusts a beta coefficient for regression dilution using a calculated value of lambda.
}
\details{
XXX THIS IS ONLY FOR COX MODELS AT THE MOMENT
}
