% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_helpers.R
\name{adjust_se}
\alias{adjust_se}
\title{Adjust standard errors on beta coefficients}
\usage{
adjust_se(beta, beta_se, lambda, var_lambda)
}
\arguments{
\item{beta}{Coefficient from epidemiological model to be adjusted}

\item{beta_se}{Standard error on beta coefficient from epidemiological model}

\item{lambda}{Regression dilution coefficient to adjust}

\item{var_lambda}{Variance of regression dilution coefficient estimate}
}
\value{
Regression-dilution adjusted beta coefficients
}
\description{
XXX ADD DESCRIPTION (arguments can be single values or vectors)
}
