% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_helpers.R
\name{adjust_se_oneval}
\alias{adjust_se_oneval}
\title{Adjust standard error on a beta coefficient for regression dilution}
\usage{
adjust_se_oneval(beta, beta_se, lambda, var_lambda)
}
\arguments{
\item{beta}{Coefficient from epidemiological model to be adjusted}

\item{beta_se}{Standard error on beta coefficient from epidemiological model}

\item{lambda}{Regression dilution coefficient to adjust}

\item{var_lambda}{Variance of regression dilution coefficient estimate}
}
\value{
Regression-dilution adjusted standard error of coefficient
}
\description{
XXX ADD DESCRIPTION HERE
}
