% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_reps.R
\name{calc_mean_reps}
\alias{calc_mean_reps}
\title{Calculate mean of repeats with valid data}
\usage{
calc_mean_reps(var_name, data, suffix_baseline, suffix_rep, exclusion_indic)
}
\arguments{
\item{var_name}{Name of variable to calculate lambda for}

\item{data}{Dataset containing repeated measurements data}

\item{suffix_baseline}{Suffix denoting a baseline measurement}

\item{suffix_rep}{Suffix(es) denoting a follow-up measurement}

\item{exclusion_indic}{Name of variable indicating row should be excluded for that repeat}
}
\value{
Column containing the mean of repeats with valid data
}
\description{
Calculates mean of repeats with valid data
}
