% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_adjustment.R
\name{calc_models_multi_single}
\alias{calc_models_multi_single}
\title{Calculates linear models for multivariable adjustment}
\usage{
calc_models_multi_single(
  var_names,
  data,
  suffix_baseline = "_s0",
  suffix_rep = c("_s1", "_s2", "_s3", "_s4"),
  exclusion_indic = "excluded",
  covariates = NULL
)
}
\arguments{
\item{var_names}{Name of variables to calculate models for}

\item{data}{Dataset containing repeated measurements data}

\item{suffix_baseline}{Suffix denoting a baseline measurement}

\item{suffix_rep}{Suffix(es) denoting a follow-up measurement}

\item{exclusion_indic}{Name of variable indicating row should be excluded for that repeat}

\item{covariates}{Covariates to adjust for in lambda estimation model}
}
\value{
Multivariable model
}
\description{
Calculates linear models for multivariable adjustment using repeated measurements data.
}
\details{
NOTE: uses mean value across all follow-up measurements with data
NOTE: I think this is essentially redundant with
}
