% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_on_repeat_data.R
\name{calc_precision_weights}
\alias{calc_precision_weights}
\title{Calculate precision weights}
\usage{
calc_precision_weights(data, suffix_rep, exclusion_indic, precision_weighted)
}
\arguments{
\item{data}{Dataset containing repeated measurements data}

\item{suffix_rep}{Suffix(es) denoting a follow-up measurement}

\item{exclusion_indic}{Name of variable indicating row should be excluded for that repeat (should have same suffix as indicated in \code{suffix_baseline}, \code{suffix_rep})}

\item{precision_weighted}{Indicator of whether precision weighting should be used in model for follow up based on number of repeats}
}
\value{
Vector of precision weights
}
\description{
Calculate precision weights
}
