% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_assumptions.R
\name{correlate_errors}
\alias{correlate_errors}
\title{Assesses correlation between different errors}
\usage{
correlate_errors(
  var_names,
  data,
  suffix_baseline = "_s0",
  suffix_rep = c("_s1", "_s2", "_s3", "_s4"),
  exclusion_indic = "excluded",
  covariates = NULL
)
}
\arguments{
\item{var_names}{Name of variables to calculate models for}

\item{data}{Dataset containing repeated measurements data}

\item{suffix_baseline}{Suffix denoting a baseline measurement}

\item{suffix_rep}{Suffix(es) denoting a follow-up measurement}

\item{exclusion_indic}{Name of variable indicating row should be excluded for that repeat}
}
\value{
Correlation matrix
}
\description{
Assesses correlation between different errors
}
