% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_on_repeat_data.R
\name{make_inc_ind}
\alias{make_inc_ind}
\title{Make inclusion indicator}
\usage{
make_inc_ind(data, suffix_rep, exclusion_indic)
}
\arguments{
\item{data}{Dataset containing repeated measurements data}

\item{suffix_rep}{Suffix(es) denoting a follow-up measurement}

\item{exclusion_indic}{Name of variable indicating row should be excluded for that repeat (should have same suffix as indicated in \code{suffix_baseline}, \code{suffix_rep})}
}
\value{
Add inclusion indicator for each repeat to data
}
\description{
Make inclusion indicator
}
