% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_on_repeat_data.R
\name{n_valid_repeats}
\alias{n_valid_repeats}
\title{Count valid repeats}
\usage{
n_valid_repeats(data, suffix_rep, exclusion_indic)
}
\arguments{
\item{data}{Dataset containing repeated measurements data}

\item{suffix_rep}{Suffix(es) denoting a follow-up measurement}

\item{exclusion_indic}{Name of variable indicating row should be excluded for that repeat (should have same suffix as indicated in \code{suffix_baseline}, \code{suffix_rep})}
}
\value{
Vector with number of valid repeats for each row
}
\description{
Utility function to count valid repeats
}
