% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_repeat.R
\name{pred_rep}
\alias{pred_rep}
\title{Predicts true val using reg cal framework}
\usage{
pred_rep(models_list, newdata, suffix_pred = "_pred")
}
\arguments{
\item{models_list}{Names list of models (probably output of calc_models_multi)}

\item{suffix_pred}{Suffix to append to name of predicted variable}

\item{new_data}{Dataset to predict true values for}
}
\value{
List of models for different outcomes
}
\description{
Predicts true val using reg cal framework
}
