% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_1_dp}
\alias{round_1_dp}
\title{Utility function to round to 1 d.p.}
\usage{
round_1_dp(x)
}
\arguments{
\item{x}{Object to round}
}
\value{
Input rounded to 1 dp
}
\description{
Utility function to round to 1 d.p.
}
