% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_2_dp}
\alias{round_2_dp}
\title{Utility function to round to 3 d.p.}
\usage{
round_2_dp(x)

round_2_dp(x)
}
\arguments{
\item{x}{Object to round}
}
\value{
Input rounded to 3 dp

Input rounded to 2 dp
}
\description{
Utility function to round to 3 d.p.

Utility function to round to 2 d.p.
}
