% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_3_dp}
\alias{round_3_dp}
\title{Utility function to round to 3 d.p.}
\usage{
round_3_dp(x)
}
\arguments{
\item{x}{Object to round}
}
\value{
Input rounded to 3 dp
}
\description{
Utility function to round to 3 d.p.
}
