% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_helpers.R
\name{adjust_beta}
\alias{adjust_beta}
\title{Adjust beta coefficients for regression dilution}
\usage{
adjust_beta(beta, lambda)
}
\arguments{
\item{beta}{Coefficients from epidemiological model to be adjusted}

\item{lambda}{Regression dilution coefficient}
}
\value{
Regression-dilution adjusted beta coefficients
}
\description{
Adjusts beta coefficients for regression dilution using a calculated value of lambda. Arguments can be single values or vectors. This is a vectorised version of univariable adjustment i.e. it performs univariable adjustment for multiple models at once. It does NOT perform multivariable regression calibration.
}
